+++ clan/clanMember/roles/rolesManagement.js
@@ -1,4 +1,4 @@
-import config from "../../config.json";
+import config from "../../../config.json";
 
 export function LogRolesGranting(displayName, isDiscordMemberFound, medals) {
 	if (medals == null) {
